pyproject.toml
[tool.poetry]
name = "restaurant-recommendation"
version = "0.1.0"
description = ""
authors = ["kai-yamashita <kai.yamashita@weblab.t.u-tokyo.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
